server/requirements.txt
fastapi
pandas
gitpython
uvicorn